requirements.txt
tensorflow>=2.12.0
gradio>=3.40.0
pillow
numpy
